Update JobService.java
 changes in deleteJob

+++ JobApp-BackEnd/demo/src/main/java/com/rupesh/demo/service/JobService.java
@@ -36,7 +36,10 @@ public class JobService {
 
 
     public void deleteJob(int jobId) {
-        repo.deleteById(jobId);
+        // repo.deleteById(jobId);
+        if(jobId!=0){
+            repo.deleteById(jobId);
+        }
     }
 
     public void loadJob(){
